+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/export/v2/SerializationBasedLedgerDataExporter.scala
@@ -1,12 +1,20 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.export
+package com.daml.ledger.participant.state.kvutils.export.v2
 
 import java.io.DataOutputStream
 import java.time.Instant
 import java.util.concurrent.locks.StampedLock
 
+import com.daml.ledger.participant.state.kvutils.export.{
+  InMemorySubmissionAggregator,
+  LedgerDataExporter,
+  LedgerDataWriter,
+  SubmissionAggregator,
+  SubmissionInfo,
+  WriteSet
+}
 import com.daml.ledger.validator.LedgerStateOperations.{Key, Value}
 
 /**
