+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/export/v3/ProtobufBasedLedgerDataExporter.scala
@@ -1,12 +1,21 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.export
+package com.daml.ledger.participant.state.kvutils.export.v3
 
 import java.io.OutputStream
 
 import com.daml.ledger.participant.state.kvutils.Conversions
 import com.daml.ledger.participant.state.kvutils.DamlKvutils.LedgerExportEntry
+import com.daml.ledger.participant.state.kvutils.export.{
+  Header,
+  InMemorySubmissionAggregator,
+  LedgerDataExporter,
+  LedgerDataWriter,
+  SubmissionAggregator,
+  SubmissionInfo,
+  WriteSet
+}
 import com.daml.ledger.validator.LedgerStateOperations.{Key, Value}
 
 final class ProtobufBasedLedgerDataExporter private (output: OutputStream)
